# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TestFlow AI QE API
  version: 1.0.0
  description: Authenticated Quality Engineering routes. JWT or X-API-Key required unless noted.
servers:
  - url: /api
paths:
...
  /artifacts/{projectId}:
    get:
      summary: List project artifacts
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Artifact list }
  /artifacts/{projectId}/file:
    get:
      summary: Download an owned artifact
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
        - in: query
          name: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: File }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - bearerAuth: []
  - apiKey: []